+++ packages/atomic/src/components/insight/atomic-insight-interface/store.ts
@@ -1,13 +1,20 @@
+import {InsightEngine} from '@coveo/headless/insight';
 import {
   AtomicCommonStoreData,
   createAtomicCommonStore,
-} from '../../common/interface/store';
+} from '@components/common/interface/store';
 
 export interface AtomicSvgInsightStoreData extends AtomicCommonStoreData {}
 
 export function createAtomicSvcInsightStore() {
-  return createAtomicCommonStore<AtomicSvgInsightStoreData>({
+  const commonStore = createAtomicCommonStore<AtomicSvgInsightStoreData>({
     loadingFlags: [],
     iconAssetsPath: '',
   });
+  return {
+    ...commonStore,
+    getUniqueIDFromEngine(engine: InsightEngine): string {
+      return engine.state.search.searchResponseId;
+    },
+  };
 }