+++ packages/atomic/src/components/search/atomic-breadbox/atomic-breadbox.tsx
@@ -3,7 +3,7 @@ import {
   InitializableComponent,
   BindStateToController,
   InitializeBindings,
-} from '../../../utils/initialization-utils';
+} from '@utils/initialization-utils';
 import {
   BreadcrumbManagerState,
   BreadcrumbManager,
@@ -12,14 +12,11 @@ import {
   FacetManagerState,
   buildFacetManager,
 } from '@coveo/headless';
-import {Button} from '../../common/button';
+import {Button} from '@components/common/button';
 import CloseIcon from 'coveo-styleguide/resources/icons/svg/close.svg';
-import {getFieldValueCaption} from '../../../utils/field-utils';
-import {Hidden} from '../../common/hidden';
-import {
-  FocusTarget,
-  FocusTargetController,
-} from '../../../utils/accessibility-utils';
+import {getFieldValueCaption} from '@utils/field-utils';
+import {Hidden} from '@components/common/hidden';
+import {FocusTarget, FocusTargetController} from '@utils/accessibility-utils';
 import {Bindings} from '../atomic-search-interface/atomic-search-interface';
 
 interface Breadcrumb {