+++ packages/atomic/src/components/search/search-box-suggestions/atomic-search-box-recent-queries/atomic-search-box-recent-queries.tsx
@@ -1,4 +1,4 @@
-import Clock from '../../../../images/clock.svg';
+import Clock from '@images/clock.svg';
 import {
   buildRecentQueriesList,
   HighlightUtils,
@@ -11,9 +11,9 @@ import {
   SearchBoxSuggestions,
   SearchBoxSuggestionsBindings,
 } from '../suggestions-common';
-import {once} from '../../../../utils/utils';
-import {SafeStorage, StorageItems} from '../../../../utils/local-storage-utils';
-import {cleanUpString} from '../../../../utils/string-utils';
+import {once} from '@utils/utils';
+import {SafeStorage, StorageItems} from '@utils/local-storage-utils';
+import {encodeForDomAttribute} from '@utils/string-utils';
 
 /**
  * The `atomic-search-box-recent-queries` component can be added as a child of an `atomic-search-box` component, allowing for the configuration of recent query suggestions.
@@ -156,7 +156,7 @@ export class AtomicSearchBoxRecentQueries {
   private renderItem(value: string): SearchBoxSuggestionElement {
     const query = this.bindings.searchBoxController.state.value;
     return {
-      key: `recent-${cleanUpString(value)}`,
+      key: `recent-${encodeForDomAttribute(value)}`,
       query: value,
       content: (
         <div class="flex items-center break-all">