+++ packages/atomic-angular/projects/atomic-angular/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.18.5](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.18.4...@coveo/atomic-angular@1.18.5) (2022-06-30)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
+## [1.18.4](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.18.3...@coveo/atomic-angular@1.18.4) (2022-06-30)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
+## [1.18.3](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.18.2...@coveo/atomic-angular@1.18.3) (2022-06-30)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
 ## [1.18.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.18.1...@coveo/atomic-angular@1.18.2) (2022-06-29)
 
 **Note:** Version bump only for package @coveo/atomic-angular